pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "psenv"
version = "0.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"python-dotenv==0.20.0",
	"pyaml==21.10.1",
	"boto3==1.20.49"
]

[project.optional-dependencies]
dev = [
	"setuptools==63.1.0",
	"build==0.8.0",
	"python-semantic-release==7.29.4",
	"pytest==7.1.2",
	"black==22.6.0"
]


[project.scripts]
psenv = "psenv.__main__:main"


[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.semantic_release]
version_variable = "psenv/cli/version.py:__version__"
version_toml = "pyproject.toml:project.version"
version_pattern = "README.md:VERSION {version}"
upload_to_pypi = false

[tool.black]
line_length = 120
